backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.40.0
httpx==0.26.0
pydantic==2.5.0
pdf2image==1.16.3
pillow==10.1.0
pytesseract==0.3.10
openpyxl==3.1.5
python-dotenv==1.0.0
pypdf==3.17.1
requests==2.31.0
